@@ -2,6 +2,7 @@
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:balance_me/firebase_wrapper/auth_repository.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:balance_me/global/types.dart';
 
 // Singleton that is used for sending logs to Google Analytics.
 // you should use it as "GoogleAnalytics.instance.LogSomething();"
@@ -28,7 +29,11 @@ class GoogleAnalytics {
     logEvent("$functionName has failed since pre-check failed", {"authRepository": authRepository, "user": authRepository.user, "email": authRepository.user!.email!});
   }
 
-  void logPostLoginFailed(DocumentSnapshot<Map<String, dynamic>> generalInfo) {
+  void logPostLoginFailed(DocumentSnapshot<Json> generalInfo) {
     logEvent("PostLoginFailed", {"dataExists": generalInfo.exists, "data": generalInfo.data()});
+  }
+
+  void logCategoriesForBalanceFailed(DocumentSnapshot<Json> generalInfo) {
+    logEvent("categoriesForBalanceFailed", {"dataExists": generalInfo.exists, "data": generalInfo.data()});
   }
 }
