@@ -1,29 +1,17 @@
 // ================= Balance Page =================
 import 'package:flutter/material.dart';
+import 'package:balance_me/localization/resources/resources.dart';
 import 'package:balance_me/firebase_wrapper/auth_repository.dart';
 import 'package:balance_me/firebase_wrapper/storage_repository.dart';
-import 'package:balance_me/common_models/category_model.dart';
-import 'package:balance_me/global/types.dart';
+import 'package:balance_me/pages/balance/balance_model.dart';
+import 'package:balance_me/widgets/ring_pie_chart.dart';
 
 class BalancePage extends StatefulWidget {
-  BalancePage(this._authRepository, this._userStorage, {Key? key}) : super(key: key) {
-    if (_authRepository.status == Status.Authenticated) {
-      pullBalanceData();
-      parseBalanceData();
-    }
-  }
-
-  void pullBalanceData() {
-    // TODO
-  }
-
-  void parseBalanceData() {
-    // TODO
-  }
+  BalancePage(this._authRepository, this._userStorage, this._balanceModel, {Key? key}) : super(key: key);
 
   final AuthRepository _authRepository;
   final UserStorage _userStorage;
-  List<Category> _categories = [];
+  final BalanceModel _balanceModel;
 
   @override
   _BalancePageState createState() => _BalancePageState();
@@ -32,10 +20,14 @@ class BalancePage extends StatefulWidget {
 class _BalancePageState extends State<BalancePage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        
-      ),
+    return Column(
+      children: [
+        Row(
+          children: [
+            RingPieChart([], Languages.of(context)!.income),
+          ],
+        )
+      ],
     );
   }
 }
